pyproject.toml
[tool.poetry]
name = "contextualizer-harness"
version = "0.1.0"
description = "Self-training contextualizer with MLM for Qwen models"
authors = ["Larry <larry@example.com>"]  # Update with your info
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.2"
transformers = "^4.44.0"
datasets = "^2.20.0"
tqdm = "^4.66"
numpy = "^1.24"
pyyaml = "^6.0"
safetensors = "^0.4"
pandas = "^2.0"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
ruff = "^0.1.0"
ipython = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
